Fix task priority type

+++ Lab10/task.h
@@ -3,7 +3,7 @@
 
 typedef struct task {
     char* task_name;
-    size_t priority;
+    int priority;
 } task_t;
 
 #endif /* TASK_H */
